sha2 ( =64 -bit word sizes ) related common functions in seperate namespace
